added force link playbook parameter, some minor refactoring
